Refactor Compile* reports
Work towards common interface for running reports -
Report::Whatever.new(params).run

- use consistent interface in CommitmentReplacements,
  UncommittedHoldings
- merge CompileCostReports -> CostReport (could consider just calling it Report::Cost?)
- merge CompileMemberCountsReport, MemberCountsReport -> MemberCounts
- merge CompileEstimatedICCosts, EstimateIC -> Estimate
- eliminate some redundant tests from member counts report
