+++ lib/reports/estimate.rb
@@ -6,12 +6,12 @@ require "reports/cost_report"
 
 module Reports
   # Generate IC estimate from a list of OCNS
-  class EstimateIC
-    attr_accessor :ocns, :h_share_total, :num_ocns_matched, :num_items_matched, :num_items_pd,
+  class Estimate
+    attr_accessor :ocns, :ocn_file, :h_share_total, :num_ocns_matched, :num_items_matched, :num_items_pd,
       :num_items_ic, :clusters_seen, :marker
 
-    def initialize(ocns, batch_size = 100_000)
-      @ocns = ocns.uniq
+    def initialize(ocn_file = nil, batch_size = 100_000)
+      @ocn_file = ocn_file
       @h_share_total = 0
       @num_ocns_matched = 0
       @num_items_matched = 0
@@ -25,7 +25,7 @@ module Reports
       @cost_report ||= CostReport.new
     end
 
-    def run
+    def find_matching_ocns(ocns = @ocns)
       ocns.each do |ocn|
         marker.incr
         cluster = Cluster.find_by(ocns: ocn.to_i,
@@ -41,6 +41,27 @@ module Reports
         marker.on_batch { |m| Services.logger.info m.batch_line }
       end
       Services.logger.info marker.final_line
+    end
+
+    def run(output_filename = report_file(ocn_file))
+      @ocns = File.open(ocn_file).map(&:to_i).uniq
+
+      Services.logger.info "Target Cost: #{cost_report.target_cost}"
+      Services.logger.info "Cost per volume: #{cost_report.cost_per_volume}"
+      Services.logger.info "Starting #{Pathname.new(__FILE__).basename}. Batches of #{ppnum marker.batch_size}"
+
+      find_matching_ocns(ocns)
+
+      File.open(output_filename, "w") do |fh|
+        fh.puts [
+          "Total Estimated IC Cost: $#{total_estimated_ic_cost.round(2)}",
+          "In all, we received #{ocns.count} distinct OCLC numbers.",
+          "Of those distinct OCLC numbers, #{num_ocns_matched} (#{pct_ocns_matched.round(1)}%) match items in",
+          "HathiTrust, corresponding to #{num_items_matched} HathiTrust items.",
+          "Of those items, #{num_items_pd} (#{pct_items_pd.round(1)}%) are in the public domain,",
+          "#{num_items_ic} (#{pct_items_ic.round(1)}%) are in copyright."
+        ].join("\n")
+      end
     end
 
     def pct_ocns_matched
@@ -77,6 +98,11 @@ module Reports
         overlap.matching_members << "prospective_member"
         @h_share_total += overlap.h_share("prospective_member")
       end
+    end
+
+    def report_file(ocn_file)
+      FileUtils.mkdir_p(Settings.estimates_path)
+      File.join(Settings.estimates_path, File.basename(ocn_file, ".txt") + "-estimate-#{Date.today}.txt")
     end
   end
 end
