+++ spec/reports/estimate_spec.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
 require "spec_helper"
-require "reports/estimate_ic"
+require "reports/estimate"
 require "clustering/cluster_ht_item"
 
-RSpec.describe Reports::EstimateIC do
+RSpec.describe Reports::Estimate do
   let(:ht_allow) { build(:ht_item, access: "allow") }
   let(:ht_deny) { build(:ht_item, access: "deny") }
   let(:ocns) { [1, 2, ht_allow.ocns, ht_deny.ocns].flatten }
-  let(:rpt) { described_class.new(ocns) }
+  let(:rpt) { described_class.new }
 
   before(:each) do
     @old_cost = Settings.target_cost
@@ -29,19 +29,19 @@ RSpec.describe Reports::EstimateIC do
     end
   end
 
-  describe "#run" do
+  describe "#find_matching_ocns" do
     it "sets num_items_pd" do
-      rpt.run
+      rpt.find_matching_ocns(ocns)
       expect(rpt.num_items_pd).to eq(1)
     end
 
     it "sets num_items_ic" do
-      rpt.run
+      rpt.find_matching_ocns(ocns)
       expect(rpt.num_items_ic).to eq(1)
     end
 
     it "compiles h_share_total" do
-      rpt.run
+      rpt.find_matching_ocns(ocns)
       # the contributor gets the other half of the IC item
       expect(rpt.h_share_total).to eq(0.5)
     end
@@ -49,7 +49,7 @@ RSpec.describe Reports::EstimateIC do
 
   describe "#total_estimate_ic_cost" do
     it "calculates total from h_share_total and cost_per_volume" do
-      rpt.run
+      rpt.find_matching_ocns(ocns)
       expect(rpt.total_estimated_ic_cost).to eq(5)
     end
   end
